Statistics: * Rename DistributionCharacteristics to DistributionStats * Update all usages of DistributionCharacteristics
